feat: split telegram pages
