+++ src/modules/telegram/components/TelegramGroupList.jsx
@@ -1,0 +1,42 @@
+import React from "react";
+
+export default function TelegramGroupList({ groups = [], onRefreshAdmins }) {
+    return (
+        <table className="table">
+            <thead>
+                <tr>
+                    <th>Назва</th>
+                    <th>chat_id</th>
+                    <th>Статус</th>
+                    <th>Дата прив'язки</th>
+                    <th></th>
+                </tr>
+            </thead>
+            <tbody>
+                {groups.map((g) => (
+                    <tr key={g.id}>
+                        <td>{g.title}</td>
+                        <td>{g.chat_id}</td>
+                        <td>{g.is_active ? "активна" : ""}</td>
+                        <td>{g.linked_at}</td>
+                        <td>
+                            {onRefreshAdmins && (
+                                <button
+                                    className="btn"
+                                    onClick={() => onRefreshAdmins(g.id)}
+                                >
+                                    Оновити адмінів
+                                </button>
+                            )}
+                        </td>
+                    </tr>
+                ))}
+                {groups.length === 0 && (
+                    <tr>
+                        <td colSpan="5">Немає груп</td>
+                    </tr>
+                )}
+            </tbody>
+        </table>
+    );
+}
